Extract SelectExpression construction out to RelationalQueryableMethodTranslatingEV
